VST2: Fix creation of a rack when one of the parameters is a NaN
On some quirky Windows VST2 (old SWAM instruments) or old Linux ones (InsertPizHere plugins), a NaN parameter value would be sent.
This throws an error in the frontend code and prevent carla from creating the corresponding rack.
We workaround that by just returning early from PixmapDialsetValue() if the value is a NaN

Fixes #1143