Add disconnect-all-of-group menu item to patchbay
If one group has multiple connections to another group an extra menu
item is added to its disconnect context menu, which allows the user to
disconnect it all at once. The entry is added at the bottom of the
disconnect menu in the form of "Groupname (all x ports)". This comes in
very handy when the user wants to disconnect a soundcard from their
current application.

PortContextMenuDisconnect() now also accepts a list via
setData()/data(), so we can disconnect multiple ports at once.