Fix canvaspreview out-of-bounds panning for patchbay
Using the small canvaspreview to pan the patchbay resulted in an
exception when panning out of bounds of the window. In this case Carla
tried to create a QPoint with two floats, for which qt has no
constructor. This results in the panning to hang, as the event is
"handled" with an exception.

Casting these two values to int lets qt create the QPoint
and with that properly set the new cursor position, so panning can be
done in all edges again, without having to be precise with the cursor
positioning.

This is probably a left-over from a python2 to python3 migration, as
python did integer divison by default.